causal_ai_chatbot/requirements_chatbot.txt
# Requirements for Causal AI Agent Chatbot

# FastAPI and WebSocket
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Local LLM client
requests>=2.31.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0  # For Excel file support

# TRAM-DAG dependencies (required for tramdag to work)
# Note: tramdag itself is in the parent directory, but these are its dependencies
torch>=2.0.0
matplotlib>=3.5.0
scipy>=1.9.0
scikit-learn>=1.0.0
joblib>=1.1.0
statsmodels>=0.13.0

# R integration
# rpy2>=3.5.0  # Optional, for direct R integration

# File handling
pathlib2>=2.3.7  # For Python < 3.6

# JSON and serialization
pydantic>=2.0.0

# PDF generation
reportlab>=4.0.0
Pillow>=10.0.0  # For image handling in PDFs
networkx>=3.0  # For DAG visualization

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # Required by FastAPI TestClient
